Instituted Changes from User Waveform
Comments are: 

Confirm this is a HNT DAO HIP to be voted with veHNT
What makes the score "a more efficient equation"? Simplified as mentioned later yes
Clarify emissions with HNT emissions everywhere its mentioned
Explain "zombie network"
However, with the implementation of HIP 19, the Helium network has established additional protections that make the inclusion of the onboarding fee in the A score calculation unnecessary...... Which are?
Stakeholder 1 is a veHNT Delegator - Delegators receive emissions
Stakeholder 3 is a veHNT Staker
Is there a further stakeholder that is a token investor?
"This could cause temporary disruptions or confusion." I think confusion is guaranteed for some.  
Key reason 3 is very disputable - isnt the disscusison going to be about this claim?
Alternatives 1: Why is the current calc more complex its just multiplying 2 numbers and one is a fixed constant